OPNET-351: Extend infra CR to store VIPs and MachineNetwork
In order to enable work around mutable VIPs we are extending the
Infrastructure CR to hold VIPs in the spec and not only in the status.

In order to be able to perform proper validations (which currently only
happen in o/installer as VIPs are immutable) we are storing machine
networks alongside.

Implements: OPNET-351
Contributes-to: OPNET-80
Contributes-to: OPNET-340